Reuse existing OpenShift cluster for 'osbs-box.py up --no-cleanup'
Use the --use-existing-config and --host-data-dir options to
'oc cluster up' to restart with the same openshift cluster configuration
for 'osbs-box.py up --no-cleanup' - the old behavior of keeping the
data volumes but recreating the openshift config confused the startup
scripts.

(Problem: the --host-data-dir directory is created as root, which causes
problems when we do want to start from scratch. Work around this with
'sudo rm -rf'.)