Switch build image to owtaylor/fedora-buildroot
owtaylor/fedora-buildroot is a modified version of the buildroot which
builds git branches of atomic-reactor/osbs-client and skopeo with
Flatpak/OCI support.

It also has a change to read entries from the reactor config secret
and add them to /etc/hosts - this is needed to work around certification
verification problems with Koji and the desire of PDC/Django to have
a real hostname.